auth and boards.ts files are moved inside services folder
